XORP/Win32 merge back to mainline.  Use 32-bit-wide interface indexes.  Add XORP_INT_CAST() where needed.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6782 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/mld6igmp/mld6igmp_config.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/mld6igmp/mld6igmp_config.cc,v 1.7 2005/06/01 09:34:00 pavlin Exp $"
+#ident "$XORP: xorp/mld6igmp/mld6igmp_config.cc,v 1.8 2005/06/03 19:04:43 pavlin Exp $"
 
 
 //
@@ -109,7 +109,7 @@ Mld6igmpNode::set_config_all_vifs_done(string& error_msg)
     //
     // Remove vifs that don't exist anymore
     //
-    for (uint16_t i = 0; i < maxvifs(); i++) {
+    for (uint32_t i = 0; i < maxvifs(); i++) {
 	Vif* node_vif = vif_find_by_vif_index(i);
 	if (node_vif == NULL)
 	    continue;
